Feat: use hashcode to identify each todo and project
Project class and Todo class have an Id member variables to keep track of the unique identity of each object of these two class. Functions that modify todos and projects now receive an id as a parameter instead of an actual object of that class. This makes it more efficient and easier to associate DOM elements to the corresponding object.