Feat: add check Todo feature
Add checkbox element to HTML and add event listener to it. Add checkTodo function in ProjectManager module. When the event is triggered, the todo will be removed from uncheckedTodos and added to checkedTodos inside that project. The todo will no longer be displayed. The check variable of that todo will be set to true.