Feat: add event listener to todo element
Add description and date into the todo element and set the display to be none by default. When user clicks on the todo element itself (except checkbox or btns), it will trigger the event listener to display more details that is hidden by default. When the user clicks on it again when the details are currently shown, the details will be hidden again.