Fix: currProject still points to one of the projects in ProjectManager
When reading from local storage, the code used to create a whole new Project object for currProject, which makes currProject points to somewhere different from that particular project in projects. Therefore, when user make changes to the current project, the change won't be reflected in the local storage as well as the projects stored in ProjectManager.